# Optimize strict generic signature checking performance