This PR optimizes the performance of strict generic signature checking introduced in #16368. When a generic class or interface is instantiated, each generic method in the class or interface is instantiated with a fresh set of cloned type parameters (which we need to handle scenarios where different generations of the same type parameter are in scope). This leads to a lot of new type identities, and potentially a lot of work comparing those identities in strict generic checking mode. As an optimization we now base signature comparisons on "canonical instantiations" that uses a single type identity for all unconstrained type parameters.

Projects that make heavy use of generic methods see dramatically improved type check times. For example:

* https://github.com/falsandtru/spica: From 4.24s to 1.83s.
* https://github.com/arturkulig/immview: From 18.25s to 1.35s.
* https://github.com/emmanueltouzery/prelude.ts: From 21.34s to 0.93s.

Only changes in RWC baselines are missing error elaborations due to improved sharing.

Fixes #18257.