@arturkulig 2017-09-21
sorry taking so long.

It works! 🎉 

with **immview** everything was ok even without that fix, but just changing types to extended interfaces, but I have other, commercial private project in TS and I run into the same issue where type checking whole project was impossible anymore and the same trick as with immview did not work. So as you suggested today I installed latest version from npm (2.6.0-dev-something) and it completes!
I no longer have compile it with increasing `--max-old-space-size` on node to run type checking too, but I'm going to leave this note here for people that would like to know cheap and very unreliable workaround for 2.5